Add target stack validate to ensure either stack_name or stack_id is passed as options Modify stack update/create/delete utilities to use stack_progress class to check the status Modify delete_stack utility to use stack_id rather than stack_name (aws_sdk behaves differently depending on value passed)

+++ lib/aws_helpers/actions/cloud_formation/stack_delete.rb
@@ -1,3 +1,6 @@
+require_relative 'stack_exists'
+require_relative 'stack_progress'
+
 module AwsHelpers
   module Actions
     module CloudFormation
@@ -9,13 +12,17 @@ module AwsHelpers
           @stack_name = stack_name
           @options = options
           @stdout = options[:stdout] || $stdout
+          @stack_exists = StackExists.new(@config, @stack_name)
         end
 
         def execute
           client = @config.aws_cloud_formation_client
           @stdout.puts "Deleting #{@stack_name}"
+          return unless @stack_exists.execute
+          response = client.describe_stacks(stack_name: @stack_name).stacks.first
+          @options[:stack_id] = response.stack_id
           client.delete_stack(stack_name: @stack_name)
-          AwsHelpers::Actions::CloudFormation::PollStackStatus.new(@config, @stack_name, @options).execute
+          AwsHelpers::Actions::CloudFormation::StackProgress.new(@config, @options).execute
         end
 
       end
